server/requirements.txt
pickle
django
djangorestframework
djangocorsheaders